+++ zetaclient/bitcoin/bitcoin_client_rpc_test.go
@@ -52,7 +52,7 @@ func (suite *BitcoinClientTestSuite) SetupTest() {
 
 	btc := client.rpcClient
 
-	_, err = btc.CreateWallet("smoketest")
+	_, err = btc.CreateWallet("e2e")
 	suite.Require().NoError(err)
 	addr, err := btc.GetNewAddress("test")
 	suite.Require().NoError(err)